git {ref,}log calls: pass --no-show-signature
If the `git` configuration `log.showSignature` is set to `true`, then
some `git log` and `git reflog` invocations performed by `git-machete`
return GPG signature output that causes `git-machete` to fail. The
failures appear to occur because GPG signatures are assumed not to
appear in the output of those commands.

This commit resolves that failure mode by passing the
`--no-show-signature` flag to all `git log` and `git reflog`
queries. (Passing that flag to `git reflog` queries is necessary
because `git reflog` is porcelain for `git log` plus some additional
flags.)

Signed-off-by: Geoffrey M. Oxberry <goxberry@gmail.com>
Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>